requirements.txt
groq==1.0.0
numpy==2.4.2
pandas==3.0.0
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.54.0
